requirements.txt
Django==4.2
djangorestframework==3.15.2
drf-spectacular==0.27.2
gunicorn==20.1.0
djoser
drf-nested-routers